+++ frontend/src/app/(routes)/settings/feegrant/components/DialogFeegrantDetails.tsx
